+++ packages/core/tests/Unit/FieldTypes/TranslatedTextTest.php
@@ -35,6 +35,25 @@ class TranslatedTextTest extends TestCase
     }
 
     /** @test */
+    public function can_json_encode_fieldtype()
+    {
+        $data = [
+            'en' => 'Blue',
+            'fr' => 'Bleu',
+        ];
+
+        $field = new TranslatedText(collect([
+            'en' => new Text('Blue'),
+            'fr' => new Text('Bleu'),
+        ]));
+
+        $this->assertSame(
+            json_encode($data),
+            json_encode($field)
+        );
+    }
+
+    /** @test */
     public function check_does_not_allow_non_text_field_types()
     {
         $this->assertTrue(true);